move components directory outside
