+++ simulator/frontend/src/components/Banners/ErrorBanner.js
