+++ simulator/frontend/src/components/Banners/SuccessBanner.js
