+++ simulator/frontend/src/components/GenericWarning.js
