+++ simulator/frontend/src/components/LoadingSpinner.js
