+++ simulator/frontend/src/components/classCodeDialog.js
