+++ simulator/frontend/src/components/progressBar.js
